+++ src/types/zeta.ts
@@ -6125,6 +6125,9 @@ export type Zeta = {
           },
           {
             "name": "ImmediateOrCancel"
+          },
+          {
+            "name": "PostOnlySlide"
           }
         ]
       }
@@ -13338,6 +13341,9 @@ export const IDL: Zeta = {
           },
           {
             "name": "ImmediateOrCancel"
+          },
+          {
+            "name": "PostOnlySlide"
           }
         ]
       }